Takeover section
